+++ FrontEnd/src/pages/AdminPage/UserProfilesTable/PaginationButtons.jsx
@@ -1,6 +1,6 @@
 import PropTypes from 'prop-types';
 import css from './PaginationButtons.module.css';
-import { PAGE_SIZES_SELECTOR } from '../AdminPage/constants';
+import { PAGE_SIZES_SELECTOR } from '../constants';
 
 function PaginationButtons({ currentPage, totalPages, onPageChange, pageSize, onPageSizeChange }) {
     const goToPreviousPage = () => {
