+++ FrontEnd/src/pages/AdminPage/UserProfilesTable/ProfilesTable.jsx
@@ -4,7 +4,7 @@ import { useNavigate } from 'react-router-dom';
 import PaginationButtons from './PaginationButtons';
 import axios from 'axios';
 import useSWR from 'swr';
-import { DEFAULT_PAGE_SIZE } from '../AdminPage/constants';
+import { DEFAULT_PAGE_SIZE } from '../constants';
 
 const COLUMN_NAMES = [
     'ID', 'Person', 'Position', 'Company', 'Region ID', 'Phone', 'EDRPOU', 'Adress', 'IsDeleted', 'IsApproved'
