+++ FrontEnd/src/pages/AdminPage/UserProfilesTable/UserTable.jsx
@@ -4,7 +4,7 @@ import { useNavigate } from 'react-router-dom';
 import PaginationButtons from './PaginationButtons';
 import axios from 'axios';
 import useSWR from 'swr';
-import { DEFAULT_PAGE_SIZE } from '../AdminPage/constants';
+import { DEFAULT_PAGE_SIZE } from '../constants';
 
 const COLUMN_NAMES = ['ID', 'ФІО', 'Пошта'];
