+++ FrontEnd/src/pages/Authorization/LoginPage.jsx
@@ -9,8 +9,8 @@ import validator from 'validator';
 import EyeVisible from './EyeVisible';
 import EyeInvisible from './EyeInvisible';
 import classes from './LoginPage.module.css';
-import { useAuth } from '../../hooks/';
-import checkIfStaff from '../../pages/AdminPage/checkIfStaff';
+import { useAuth } from '../../hooks';
+import checkIfStaff from '../AdminPage/checkIfStaff';
 
 const LoginContent = () => {
   const { login } = useAuth();
