+++ FrontEnd/src/pages/Authorization/SignUpInvitation.jsx
