+++ FrontEnd/src/pages/SignUp/ResendActivationForm.jsx
