+++ FrontEnd/src/pages/SignUp/SignUp/ResendActivationPage.jsx
