+++ FrontEnd/src/pages/SignUp/SignupForm/ResendActivationFormContent.jsx
