+++ FrontEnd/src/pages/SignUp/SignupForm/RulesModal.jsx
