+++ FrontEnd/src/pages/SignUp/SignupForm/Rules_text.js
