+++ FrontEnd/src/pages/SignUp/SignupForm/SignUpFormContent.jsx
@@ -5,8 +5,8 @@ import { useNavigate } from 'react-router-dom';
 import { toast } from 'react-toastify';
 import axios from 'axios';
 import { Tooltip } from 'antd';
-import EyeInvisible from '../../Authorization/EyeInvisible';
-import EyeVisible from '../../Authorization/EyeVisible';
+import EyeInvisible from '../../../pages/Authorization/EyeInvisible';
+import EyeVisible from '../../../pages/Authorization/EyeVisible';
 import styles from './SignUpFormContent.module.css';
 import PropTypes from 'prop-types';
 import {
