+++ FrontEnd/src/routes/ClientRouter.jsx
@@ -7,7 +7,7 @@ import { SWRConfig } from 'swr';
 import 'react-toastify/dist/ReactToastify.css';
 import customTheme from '../pages/CustomThemes/customTheme.js';
 
-import AuthorizationPage from '../components/Authorization/AuthorizationPage';
+import AuthorizationPage from '../pages/Authorization/AuthorizationPage';
 import CookiesPolicyComponent from '../pages/CookiesPolicyPage/CookiesPolicyComponent';
 import Footer from '../components/Footer/Footer';
 import Header from '../components/Header/Header';
@@ -17,10 +17,10 @@ import PrivacyPolicy from '../pages/PrivacyPolicyPage/PrivacyPolicyPage';
 import ProfileDetailPage from '../pages/ProfileDetail/ProfileDetailPage';
 import ProfileListPage from '../pages/ProfileList/ProfileListPage';
 import ProfilePage from '../pages/ProfilePage/ProfilePage';
-import { SignUpPage } from '../components/SignUp/SignUp/SignUpPage';
-import { SignUpModalPage } from '../components/SignUp/SignUp/SignUpModalPage';
-import { ResendActivationPage } from '../components/SignUp/SignUp/ResendActivationPage';
-import { ActivationProfilePage } from '../components/SignUp/SignUp/ActivateProfilePage';
+import { SignUpPage } from '../pages/SignUp/SignUp/SignUpPage';
+import { SignUpModalPage } from '../pages/SignUp/SignUp/SignUpModalPage';
+import { ResendActivationPage } from '../pages/SignUp/SignUp/ResendActivationPage';
+import { ActivationProfilePage } from '../pages/SignUp/SignUp/ActivateProfilePage';
 import { SendEmailRestorePasswordPage } from '../pages/RestorePassword/Pages/SendEmailRestorePasswordPage';
 import { RestorePasswordPage } from '../pages/RestorePassword/Pages/RestorePasswordPage';
 import { RestorePasswordModalPage } from '../pages/RestorePassword/Pages/RestorePasswordModalPage';
